Processing(): Use uint16_t as parameter in function maxChildrenSet()

+++ Processing.cpp
@@ -834,7 +834,7 @@ size_t Processing::mncpy(void *dest, size_t destSize, const void *src, size_t sr
 }
 
 #if !CONFIG_PROC_USE_STD_LISTS
-void Processing::maxChildrenSet(size_t cnt)
+void Processing::maxChildrenSet(uint16_t cnt)
 {
 	if (mpChildList)
 	{
